requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
mysqlclient==2.2.5
pillow==11.0.0
sqlparse==0.5.1
tzdata==2024.2
